+++ api/src/main/java/org/allaymc/api/debugshape/DebugText.java
@@ -15,6 +15,19 @@ public class DebugText extends DebugShape {
      */
     @Getter
     protected String text;
+    /**
+     * The scale of the text, which is a multiplier for the size of the text.
+     * <p>
+     * Can be {@code null}, and in that case that the scale will be set to 1 client-side.
+     */
+    protected Float scale;
+
+    /**
+     * @see #DebugText(Vector3fc, Color, String, Float)
+     */
+    public DebugText(Vector3fc position, Color color, String text) {
+        this(position, color, text, null);
+    }
 
     /**
      * Creates a DebugText with the specified position, color, and text.
@@ -23,9 +36,10 @@ public class DebugText extends DebugShape {
      * @param color    The color of the text
      * @param text     The text to display
      */
-    public DebugText(Vector3fc position, Color color, String text) {
+    public DebugText(Vector3fc position, Color color, String text, Float scale) {
         super(position, color);
         this.text = text;
+        this.scale = scale;
     }
 
     /**
@@ -35,6 +49,25 @@ public class DebugText extends DebugShape {
      */
     public void setText(String text) {
         this.text = text;
+        this.onChange();
+    }
+
+    /**
+     * Gets the scale of the text.
+     *
+     * @return the scale of the text
+     */
+    public float getScale() {
+        return scale != null ? scale : 1.0f;
+    }
+
+    /**
+     * Sets the scale of the text.
+     *
+     * @param scale the scale of the text, which is a multiplier for the size of the text
+     */
+    public void setScale(Float scale) {
+        this.scale = scale;
         this.onChange();
     }
 }