+++ CHANGELOG.md
@@ -16,12 +16,15 @@ Unless otherwise specified, any version comparison below is the comparison of th
 
 ### Added
 
+- (API) `DebugText` now support `scale` property.
 - Command nodes that the player doesn't have permission to access will not be sent to the player, and the commands will
   be resent when the specific permission changes.
 
 ### Changed
 
 ### Fixed
+
+- Fixed debug shape rendering in 1.21.120.
 
 ### Removed